WIP - updated chain sync supplier
